Initial commit. Just works.
Place into <Packages/User/>.
It's not supposed to work out-of-the-box. Playing with sublime plugins
and rtags to proof the concept for myself and check for possible
problems. It works with already running `rdm` daemon and installed `rc`
client with loaded database.

Move cursor to symbol. Ctrl+Shift+U "Find usages", "F2" "Go to
definition/declaration".